Index/WelCProxy/MediaProxy/requirements.txt
google-cloud-storage
google-cloud-functions
functions-framework
flask[async]
flask-cors